requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
validators
beautifulsoup4